Updating drugs through redux and constructing an array with a schema identical
Here are some major details that I added:
In constants.js: when defining the startDate of the drugs, I instantiated their dates only with hour and minutes (someone will need to add month and day too to calculate the next time the reminder is on). I also made an array of reminders in unsorted order to check if my function  sorts them in DayViewScreen.

In DayViewScreen.js: defining organizeDrugsByEvent() and getDrug(), taking reminders and drugs array from redux.